got storybook webworkers working

+++ packages/dcf-react/src/workers/dcfModel.worker.js
@@ -1,27 +1,32 @@
-import DCFModelSubWorker from "./dcfModel.sub.worker";
+let WorkerC = Worker;
+
+if (process.env.NODE_ENV === "development") {
+  const DCFModelSubWorker = require("./dcfModel.sub.worker");
+
+  WorkerC = DCFModelSubWorker.default;
+}
 
 const initializeSubWorker = () => {
-  return new DCFModelSubWorker("./dcfModelSub.worker.js", {
+  return new WorkerC("./dcfModel.sub.worker.js", {
     type: "module",
   });
 };
 
-let subWorker = null;
+if (typeof self !== "undefined") {
+  let subWorker = null;
 
-self.onmessage = ({ data: { cells, existingScope, currentScopes } }) => {
-  if (subWorker) {
-    subWorker.terminate();
-  }
-
-  subWorker = initializeSubWorker();
-
-  subWorker.postMessage({
-    cells,
-    existingScope,
-    currentScopes,
-  });
-
-  subWorker.onmessage = ({ data }) => {
-    self.postMessage(data);
+  self.onmessage = ({ data: { cells, existingScope, currentScopes } }) => {
+    if (subWorker) {
+      subWorker.terminate();
+    }
+    subWorker = initializeSubWorker();
+    subWorker.postMessage({
+      cells,
+      existingScope,
+      currentScopes,
+    });
+    subWorker.onmessage = ({ data }) => {
+      self.postMessage(data);
+    };
   };
-};
+}
